requirements.txt
arrow==0.15.8
asgiref==3.2.10
binaryornot==0.4.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==3.0.4
django-split-settings==1.0.0
idna==2.10
Jinja2==2.11.2
jinja2-time==0.2.0
MarkupSafe==1.1.1
poyo==0.5.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.10
